fix(shader): fix shader float -> int -> float conversion causing symbol texture pulling issues

+++ src/renderer/src/renderer/engine/shaders/modules/PullSymbolParameter.frag
@@ -1,4 +1,4 @@
-float pullSymbolParameter(int offset, int SymNum) {
+float pullSymbolParameter(int offset, float SymNum) {
   vec2 texcoord = (vec2(float(offset), SymNum) + 0.5) / u_SymbolsTextureDimensions;
   vec4 pixelValue = texture2D(u_SymbolsTexture, texcoord);
   return pixelValue.x;
